Entities for the Delivery use case.
ItemDelivery implements the Delivery interface. This allows for the addition of other types of deliveries to be added later on (may need to change the message sent to the user for these other types of deliveries). 
May need to be updated by adding more variables depending on the how the distance for the delivery will be calculated.